requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.0.3
gspread==5.11.3
oauth2client==4.1.3
cryptography==41.0.3